requirements.txt
aiofiles==0.6.0
bitcoin==1.1.42
certifi==2020.6.20
chardet==3.0.4
h11==0.9.0
httpcore==0.11.1
httptools==0.1.1
httpx==0.15.4
idna==2.10
multidict==5.0.0
persist-queue==0.5.0
python-dotenv==0.15.0
requests==2.24.0
rfc3986==1.4.0
sanic==20.9.1
sniffio==1.2.0
ujson==4.0.1
urllib3==1.25.11
uvloop==0.14.0
websockets==8.1
